Flatten st_common.h
