+++ config/splat.pspeu.stno3.yaml
@@ -83,7 +83,7 @@ segments:
       - [0x24AD8, c, no3/water_effects]
       - [0x27B48, c, no3/popup]
       - [0x28140, c, no3/e_room_fg]
-      - [0x28258, c, no3_psp/st_common]
+      - [0x28258, c, no3/st_common]
       - [0x29A08, c, no3_psp/working_on]
       - [0x2A680, data]
       - [0x33BE0, .data, no3_psp/e_init]
