--- src/st/adjust_value_within_threshold.h
@@ -1,23 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-u8 AdjustValueWithinThreshold(u8 threshold, u8 currentValue, u8 targetValue) {
-    u8 absoluteDifference;
-    s8 relativeDifference = targetValue - currentValue;
-
-    if (relativeDifference < 0) {
-        absoluteDifference = -relativeDifference;
-    } else {
-        absoluteDifference = relativeDifference;
-    }
-
-    if (absoluteDifference > threshold) {
-        if (relativeDifference < 0) {
-            absoluteDifference = currentValue - threshold;
-        } else {
-            absoluteDifference = currentValue + threshold;
-        }
-
-        return absoluteDifference;
-    }
-
-    return targetValue;
-}
