--- src/st/check_field_collision.h
@@ -1,35 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-void CheckFieldCollision(s16* hitSensors, s16 sensorCount) {
-    Collider collider;
-    s32 velocityX;
-    s16 i;
-    s16 x;
-    s16 y;
-
-    velocityX = g_CurrentEntity->velocityX;
-    if (velocityX == 0) {
-        return;
-    }
-
-    x = g_CurrentEntity->posX.i.hi;
-    y = g_CurrentEntity->posY.i.hi;
-    for (i = 0; i < sensorCount; i++) {
-        if (velocityX < 0) {
-            x += *hitSensors++;
-        } else {
-            x -= *hitSensors++;
-        }
-
-        y += *hitSensors++;
-        g_api.CheckCollision(x, y, &collider, 0);
-        if (collider.effects & EFFECT_UNK_0002 &&
-            (!(collider.effects & EFFECT_UNK_8000) || i)) {
-            if (velocityX < 0) {
-                g_CurrentEntity->posX.i.hi += collider.unk1C;
-            } else {
-                g_CurrentEntity->posX.i.hi += collider.unk14;
-            }
-            break;
-        }
-    }
-}
