--- src/st/entity_dummy.h
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-void EntityDummy(Entity* arg0) {
-    if (!arg0->step) {
-        arg0->step++;
-    }
-}
