--- src/st/get_angle_between_entities.h
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-u16 GetAngleBetweenEntities(Entity* a, Entity* b) {
-    s32 diffX = b->posX.i.hi - a->posX.i.hi;
-    s32 diffY = b->posY.i.hi - a->posY.i.hi;
-    return ratan2(diffY, diffX);
-}
