--- src/st/get_angle_point_to_entity_shifted.h
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-u8 GetAnglePointToEntityShifted(s32 x, s32 y) {
-    s32 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
-    s32 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
-    return Ratan2Shifted(diffX, diffY);
-}
