--- src/st/get_sine.h
@@ -1,2 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-s16 GetSine(u8 arg0) { return g_SineTable[arg0]; }
