--- src/st/move_entity.h
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-void MoveEntity(void) {
-    g_CurrentEntity->posX.val += g_CurrentEntity->velocityX;
-    g_CurrentEntity->posY.val += g_CurrentEntity->velocityY;
-}
