--- src/st/no3_psp/st_common.c
@@ -1,46 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "common.h"
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", DestroyEntity);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", PreventEntityFromRespawning);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", AnimateEntity);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", GetDistanceToPlayerX);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", GetDistanceToPlayerY);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", GetSideToPlayer);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", MoveEntity);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", FallEntity);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", UnkCollisionFunc3);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", UnkCollisionFunc2);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", AllocEntity);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", UnkEntityFunc0);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", GetAngleBetweenEntities);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", GetNormalizedAngle);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", SetStep);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", SetSubStep);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", InitializeEntity);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", EntityDummy);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", UnkCollisionFunc);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", CheckFieldCollision);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", GetPlayerCollisionWith);
-
-INCLUDE_ASM("st/no3_psp/psp/no3_psp/st_common", ReplaceBreakableWithItemDrop);
