--- src/st/ratan2.h
@@ -1,2 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-u16 Ratan2(s16 x, s16 y) { return ratan2(y, x); }
