--- src/st/set_step.h
@@ -1,7 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-void SetStep(u8 step) {
-    g_CurrentEntity->step = step;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->pose = 0;
-    g_CurrentEntity->poseTimer = 0;
-}
