--- src/st/unk_collision_func.h
@@ -1,31 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-s32 UnkCollisionFunc(s16* hitSensors, s16 sensorCount) {
-    Collider collider;
-    s32 velocityX;
-    s16 i;
-    s16 x;
-    s16 y;
-
-    velocityX = g_CurrentEntity->velocityX;
-    if (velocityX != 0) {
-        x = g_CurrentEntity->posX.i.hi;
-        y = g_CurrentEntity->posY.i.hi;
-        for (i = 0; i < sensorCount; i++) {
-            if (velocityX < 0) {
-                x += *hitSensors++;
-            } else {
-                x -= *hitSensors++;
-            }
-
-            y += *hitSensors++;
-            g_api.CheckCollision(x, y, &collider, 0);
-            if (collider.effects & EFFECT_UNK_0002 &&
-                ((!(collider.effects & EFFECT_UNK_8000)) || i)) {
-                return 2;
-            }
-        }
-        return 0;
-    }
-
-    // implicit return
-}
